refactor(filmpage): add film content component, clean up after lint
